refactor: reorganize frontend module structure
