+++ OcchioOnniveggente/src/frontend/state.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from dataclasses import dataclass, field
 from typing import Any
 
-from .conversation import ConversationManager
+from ..conversation import ConversationManager
 
 
 @dataclass
