+++ tests/test_ui_state_chat.py
@@ -3,7 +3,7 @@ from pathlib import Path
 
 sys.path.append(str(Path(__file__).resolve().parents[1]))
 
-from OcchioOnniveggente.src.ui_state import UIState, apply_to_chat
+from OcchioOnniveggente.src.frontend.state import UIState, apply_to_chat
 
 
 def test_apply_to_chat_creates_conversation():
